+++ src/view/ItemAnimations/ItemAnimationsApp.ts
@@ -1,6 +1,6 @@
-import { SvelteApplication } from '@typhonjs-fvtt/runtime/svelte/application'
-
 import type { CombinedSvelteApplicationOptions, ConstructorApplicationOptions } from 'src/extensions'
+import { SvelteApplication } from '@typhonjs-fvtt/runtime/svelte/application'
+import { kofiButton } from 'src/utils'
 import BasicAppShell from './ItemAnimationsShell.svelte'
 
 interface ItemAnimationsOptions {
@@ -40,14 +40,7 @@ export default class ItemAnimationsApp extends SvelteApplication {
 
 	override _getHeaderButtons() {
 		const buttons = super._getHeaderButtons()
-		buttons.unshift({
-			icon: 'fas fa-mug-hot ko-fi',
-			class: '',
-			label: 'pf2e-graphics.support',
-			onclick: () => {
-				window.open('https://ko-fi.com/mrvauxs', '_blank')
-			},
-		})
+		kofiButton(buttons)
 		return buttons
 	}
 }