+++ src/view/index.ts
@@ -1,2 +1,3 @@
 import './ActorAnimations'
 import './ItemAnimations'
+import './UserAnimations'